qty-update

+++ models/product.js
@@ -11,7 +11,7 @@ module.exports = (sequelize, DataTypes) => {
     },
     
     price_per_unit: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DOUBLE,
       allowNull: false,
     },
     discount: {
@@ -23,7 +23,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     stock_quantity: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       allowNull: false,
     },
